requirements.txt
requests
pandas
beautifulsoup4
lxml
tqdm
python-dateutil
PyYAML
pydantic